Add JsonProperty annotations to ErrorCode constructor.
The ErrorCode class cannot be deserialised from a JSON response currently as
it has no default constructor, and simply adding a default constructor
wouldn't work with the final fields. Add the @JsonProperty annotation to the
actual constructor so that Jackson can create instances of this.

This is needed for consumers of endpoints returning instances of HealthStatus
as HealthStatus contains an ErrorCode field.